# node/coinstacks/polygon/api/src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"TokenBalance": {
				"description": "Contains info about a token including balance for an address",
				"properties": {
					"contract": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"id": {
						"type": "string",
						"description": "nft or multi token id"
					},
					"balance": {
						"type": "string"
					}
				},
				"required": [
					"contract",
					"decimals",
					"name",
					"symbol",
					"type",
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Account": {
				"description": "Contains info about an EVM account",
				"properties": {
					"balance": {
						"type": "string"
					},
					"unconfirmedBalance": {
						"type": "string"
					},
					"pubkey": {
						"type": "string"
					},
					"nonce": {
						"type": "number",
						"format": "double"
					},
					"tokens": {
						"items": {
							"$ref": "#/components/schemas/TokenBalance"
						},
						"type": "array"
					}
				},
				"required": [
					"balance",
					"unconfirmedBalance",
					"pubkey",
					"nonce",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BadRequestError": {
				"description": "Contains info about a 400 Bad Request response",
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationError": {
				"description": "Contains info about a 422 Validation Error response",
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternalServerError": {
				"description": "Contains info about a 500 Internal Server Error response",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenTransfer": {
				"description": "Contains info about a token including transfer details",
				"properties": {
					"contract": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"id": {
						"type": "string",
						"description": "nft or multi token id"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"contract",
					"decimals",
					"name",
					"symbol",
					"type",
					"from",
					"to",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InternalTx": {
				"description": "Contains info about an EVM internal transaction",
				"properties": {
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"from",
					"to",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tx": {
				"description": "Contains info about an EVM transaction",
				"properties": {
					"txid": {
						"type": "string"
					},
					"blockHash": {
						"type": "string"
					},
					"blockHeight": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"confirmations": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					},
					"fee": {
						"type": "string"
					},
					"gasLimit": {
						"type": "string"
					},
					"gasUsed": {
						"type": "string"
					},
					"gasPrice": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"inputData": {
						"type": "string"
					},
					"tokenTransfers": {
						"items": {
							"$ref": "#/components/schemas/TokenTransfer"
						},
						"type": "array"
					},
					"internalTxs": {
						"items": {
							"$ref": "#/components/schemas/InternalTx"
						},
						"type": "array"
					}
				},
				"required": [
					"txid",
					"blockHeight",
					"timestamp",
					"from",
					"to",
					"confirmations",
					"value",
					"fee",
					"gasLimit",
					"gasPrice",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseTxHistory_Tx_": {
				"description": "Contains paginated base transaction history details",
				"properties": {
					"cursor": {
						"type": "string"
					},
					"pubkey": {
						"type": "string"
					},
					"txs": {
						"items": {
							"$ref": "#/components/schemas/Tx"
						},
						"type": "array"
					}
				},
				"required": [
					"pubkey",
					"txs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TxHistory": {
				"$ref": "#/components/schemas/BaseTxHistory_Tx_",
				"description": "Contains info about EVM transaction history"
			},
			"GasEstimate": {
				"description": "Contains info about estimated gas cost of a transaction",
				"properties": {
					"gasLimit": {
						"type": "string"
					}
				},
				"required": [
					"gasLimit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Fees": {
				"description": "Contains info about legacy and/or EIP-1559 fees",
				"properties": {
					"gasPrice": {
						"type": "string"
					},
					"maxFeePerGas": {
						"type": "string"
					},
					"maxPriorityFeePerGas": {
						"type": "string"
					}
				},
				"required": [
					"gasPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GasFees": {
				"description": "Contains info about current recommended fees to use in a transaction.\nEstimates for slow, average and fast confirmation speeds provided as well.",
				"properties": {
					"gasPrice": {
						"type": "string"
					},
					"baseFeePerGas": {
						"type": "string"
					},
					"maxPriorityFeePerGas": {
						"type": "string"
					},
					"slow": {
						"$ref": "#/components/schemas/Fees"
					},
					"average": {
						"$ref": "#/components/schemas/Fees"
					},
					"fast": {
						"$ref": "#/components/schemas/Fees"
					}
				},
				"required": [
					"gasPrice",
					"slow",
					"average",
					"fast"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendTxBody": {
				"description": "Contains the serialized raw transaction hex",
				"properties": {
					"hex": {
						"type": "string"
					}
				},
				"required": [
					"hex"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenMetadata": {
				"description": "Contains info about token metadata (ERC-721/ERC-1155)",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"media": {
						"properties": {
							"type": {
								"type": "string",
								"enum": [
									"image",
									"video"
								]
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"url"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"description",
					"media"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenType": {
				"type": "string",
				"enum": [
					"erc721",
					"erc1155"
				],
				"description": "Supported token types for token metadata"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@sudophunk/polygon-api",
		"version": "12.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/account/{pubkey}": {
			"get": {
				"operationId": "GetAccount",
				"responses": {
					"200": {
						"description": "account details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Account"
								},
								"examples": {
									"Example 1": {
										"value": {
											"balance": "5509799116755998875",
											"unconfirmedBalance": "0",
											"nonce": 6,
											"pubkey": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA",
											"tokens": [
												{
													"balance": "504032777",
													"contract": "0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
													"decimals": 6,
													"name": "USD Coin (PoS)",
													"symbol": "USDC",
													"type": "ERC20"
												}
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get account details by address",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [
					{
						"description": "account address",
						"in": "path",
						"name": "pubkey",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA"
					}
				]
			}
		},
		"/api/v1/account/{pubkey}/txs": {
			"get": {
				"operationId": "GetTxHistory",
				"responses": {
					"200": {
						"description": "transaction history",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TxHistory"
								},
								"examples": {
									"Example 1": {
										"value": {
											"pubkey": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA",
											"cursor": "eyJibG9ja2Jvb2tQYWdlIjozLCJleHBsb3JlclBhZ2UiOjEsImJsb2NrYm9va1R4aWQiOiIweGQwYzIwZmRlM2JmNWQ5NmNkNTdkN2E3MDJkYzAwYjI0ODNlOTRhZTQ0YTM0NjU3YTQzZGZkZTIyNDBmNDljNjYiLCJleHBsb3JlclR4aWQiOiIweGQwYzIwZmRlM2JmNWQ5NmNkNTdkN2E3MDJkYzAwYjI0ODNlOTRhZTQ0YTM0NjU3YTQzZGZkZTIyNDBmNDljNjYiLCJibG9ja0hlaWdodCI6NDE2Mjg3MjB9",
											"txs": [
												{
													"txid": "0x8207404cac0b02ed15dd2690fcfb618a7d7b0482e12345e4483fed398f073890",
													"blockHash": "0x00561abfd4c6a15fbeae39a1223b0a0ffbd9a4eb22e902c7b0e551049a33b2c0",
													"blockHeight": 41639607,
													"timestamp": 1681745623,
													"status": 1,
													"from": "0xe93685f3bBA03016F02bD1828BaDD6195988D950",
													"to": "0x75dC8e5F50C8221a82CA6aF64aF811caA983B65f",
													"confirmations": 155,
													"value": "0",
													"fee": "94583813568181480",
													"gasLimit": "1200108",
													"gasUsed": "223268",
													"gasPrice": "423633541610",
													"inputData": "0x252f7b0100000000000000000000000000000000000000000000000000000000000000700000000000000000000000009d1b1669c73b033dfe47ae5a0164ab96df25b944000000000000000000000000000000000000000000000000000000000003b920cb300a75b8df105ca88a050fa3480cd5e139b4dde3387beddfcdf2f9b31e8638cb300a75b8df105ca88a050fa3480cd5e139b4dde3387beddfcdf2f9b31e863800000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000002740000000000000000000000004d73adb72bc3dd368966edd0f0b2148401a178e200000000000190b3007045a01e4e04f14f7a4a6702c74187c5f6222033cd006d9d1b1669c73b033dfe47ae5a0164ab96df25b944000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000249f00000000000000000000000000000000000000000000000000000017d7309b7ab000000000000000000000000000000000000000000000000000000001e0aee090000000000000000000000000000000000000000000000000000000000003b1a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000462ee000000000000000000000000000000000000000000000000000000001e0f8c1100000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000143f758726e31b299afb85b3d5c8b1fec9b20b17ca0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
													"tokenTransfers": [
														{
															"contract": "0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
															"decimals": 6,
															"name": "USD Coin (PoS)",
															"symbol": "USDC",
															"type": "ERC20",
															"from": "0x1205f31718499dBf1fCa446663B532Ef87481fe1",
															"to": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA",
															"value": "504032777"
														}
													]
												}
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get transaction history by address",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [
					{
						"description": "account address",
						"in": "path",
						"name": "pubkey",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA"
					},
					{
						"description": "the cursor returned in previous query (base64 encoded json object with a 'page' property)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "page size (10 by default)",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/tx/{txid}": {
			"get": {
				"operationId": "GetTransaction",
				"responses": {
					"200": {
						"description": "transaction payload",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tx"
								},
								"examples": {
									"Example 1": {
										"value": {
											"txid": "0x8207404cac0b02ed15dd2690fcfb618a7d7b0482e12345e4483fed398f073890",
											"blockHash": "0x00561abfd4c6a15fbeae39a1223b0a0ffbd9a4eb22e902c7b0e551049a33b2c0",
											"blockHeight": 41639607,
											"timestamp": 1681745623,
											"status": 1,
											"from": "0xe93685f3bBA03016F02bD1828BaDD6195988D950",
											"to": "0x75dC8e5F50C8221a82CA6aF64aF811caA983B65f",
											"confirmations": 193,
											"value": "0",
											"fee": "94583813568181480",
											"gasLimit": "1200108",
											"gasUsed": "223268",
											"gasPrice": "423633541610",
											"inputData": "0x252f7b0100000000000000000000000000000000000000000000000000000000000000700000000000000000000000009d1b1669c73b033dfe47ae5a0164ab96df25b944000000000000000000000000000000000000000000000000000000000003b920cb300a75b8df105ca88a050fa3480cd5e139b4dde3387beddfcdf2f9b31e8638cb300a75b8df105ca88a050fa3480cd5e139b4dde3387beddfcdf2f9b31e863800000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000002740000000000000000000000004d73adb72bc3dd368966edd0f0b2148401a178e200000000000190b3007045a01e4e04f14f7a4a6702c74187c5f6222033cd006d9d1b1669c73b033dfe47ae5a0164ab96df25b944000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000249f00000000000000000000000000000000000000000000000000000017d7309b7ab000000000000000000000000000000000000000000000000000000001e0aee090000000000000000000000000000000000000000000000000000000000003b1a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000462ee000000000000000000000000000000000000000000000000000000001e0f8c1100000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000143f758726e31b299afb85b3d5c8b1fec9b20b17ca0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
											"tokenTransfers": [
												{
													"contract": "0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
													"decimals": 6,
													"name": "USD Coin (PoS)",
													"symbol": "USDC",
													"type": "ERC20",
													"from": "0x1205f31718499dBf1fCa446663B532Ef87481fe1",
													"to": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA",
													"value": "504032777"
												}
											],
											"internalTxs": []
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get transaction details",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [
					{
						"description": "transaction hash",
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x8207404cac0b02ed15dd2690fcfb618a7d7b0482e12345e4483fed398f073890"
					}
				]
			}
		},
		"/api/v1/gas/estimate": {
			"get": {
				"operationId": "EstimateGas",
				"responses": {
					"200": {
						"description": "estimated gas cost",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GasEstimate"
								},
								"examples": {
									"Example 1": {
										"value": {
											"gasLimit": "21000"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get the estimated gas cost of a transaction",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [
					{
						"description": "input data",
						"in": "query",
						"name": "data",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x"
					},
					{
						"description": "from address",
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x0000000000000000000000000000000000000000"
					},
					{
						"description": "to address",
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x3f758726E31b299Afb85b3D5C8B1fEc9b20b17cA"
					},
					{
						"description": "transaction value in wei",
						"in": "query",
						"name": "value",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "1337"
					}
				]
			}
		},
		"/api/v1/gas/fees": {
			"get": {
				"operationId": "GetGasFees",
				"responses": {
					"200": {
						"description": "current fees specified in wei",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GasFees"
								},
								"examples": {
									"Example 1": {
										"value": {
											"gasPrice": "142419538445",
											"baseFeePerGas": "112419538445",
											"maxPriorityFeePerGas": "30000000000",
											"slow": {
												"gasPrice": "131449097003",
												"maxFeePerGas": "140884315981",
												"maxPriorityFeePerGas": "29910015485"
											},
											"average": {
												"gasPrice": "172389951405",
												"maxFeePerGas": "160734779396",
												"maxPriorityFeePerGas": "49760478900"
											},
											"fast": {
												"gasPrice": "195530342545",
												"maxFeePerGas": "280530813993",
												"maxPriorityFeePerGas": "169556513497"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get the current recommended gas fees to use in a transaction\n\n* For EIP-1559 transactions, use `maxFeePerGas` and `maxPriorityFeePerGas`\n* For Legacy transactions, use `gasPrice`",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/send": {
			"post": {
				"operationId": "SendTx",
				"responses": {
					"200": {
						"description": "transaction id",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "0xb9d4ad5408f53eac8627f9ccd840ba8fb3469d55cd9cc2a11c6e049f1eef4edd"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Broadcast signed raw transaction",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "serialized raw transaction hex",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendTxBody",
								"description": "serialized raw transaction hex"
							},
							"example": {
								"hex": "0xf86c0a85046c7cfe0083016dea94d1310c1e038bc12865d3d3997275b3e4737c6302880b503be34d9fe80080269fc7eaaa9c21f59adf8ad43ed66cf5ef9ee1c317bd4d32cd65401e7aaca47cfaa0387d79c65b90be6260d09dcfb780f29dd8133b9b1ceb20b83b7e442b4bfc30cb"
							}
						}
					}
				}
			}
		},
		"/api/v1/metadata/token": {
			"get": {
				"operationId": "GetTokenMetadata",
				"responses": {
					"200": {
						"description": "token metadata",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenMetadata"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "On-chain Quest",
											"description": "",
											"media": {
												"url": "https://cdn.galxe.com/galaxy/symbiosis/a9e5b6a0-6e72-4f09-8a96-bd9ca313411f.png",
												"type": "image"
											}
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"description": "Get token metadata",
				"tags": [
					"v1"
				],
				"security": [],
				"parameters": [
					{
						"description": "contract address",
						"in": "query",
						"name": "contract",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "0x495D897BE6a6e57432C954f3C381eaBfa4699795"
					},
					{
						"description": "token identifier",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "5705"
					},
					{
						"description": "token type (erc721 or erc1155)",
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/TokenType"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}